requirements.dev.txt
django-debug-toolbar
django-debug-toolbar-template-timings
sphinx_rtd_theme